[LO-SRP] Extracted formatting operations within TextUI class into Formatter class
